move to earn config
