+++ runtime/selendra/src/config/earn.rs
@@ -1,0 +1,27 @@
+use crate::{
+	parameter_types, weights, BlockNumber, CdpTreasury, Currencies, Dex, EmergencyShutdown, Event,
+	GetNativeCurrencyId, GetStableCurrencyId, IncentivesPalletId, Permill, Runtime,
+	UnreleasedNativeVaultAccountId, MINUTES,
+};
+use runtime_common::EnsureRootOrThreeFourthsCouncil;
+
+parameter_types! {
+	pub const AccumulatePeriod: BlockNumber = MINUTES;
+	pub const EarnShareBooster: Permill = Permill::from_percent(30);
+}
+
+impl module_incentives::Config for Runtime {
+	type Event = Event;
+	type RewardsSource = UnreleasedNativeVaultAccountId;
+	type StableCurrencyId = GetStableCurrencyId;
+	type NativeCurrencyId = GetNativeCurrencyId;
+	type EarnShareBooster = EarnShareBooster;
+	type AccumulatePeriod = AccumulatePeriod;
+	type UpdateOrigin = EnsureRootOrThreeFourthsCouncil;
+	type CDPTreasury = CdpTreasury;
+	type Currency = Currencies;
+	type DEX = Dex;
+	type EmergencyShutdown = EmergencyShutdown;
+	type PalletId = IncentivesPalletId;
+	type WeightInfo = weights::module_incentives::WeightInfo<Runtime>;
+}
